Added optional init parameter "mimeType" to DocumentImpl.  If specified, and markupAware is true, the mimeType parameter overrides the usual format selection heuristics based on file extension, web server mime type and magic numbers.
Because this adds a field to DocumentImpl, any documents written to a
SerialDataStore from now on will not be readable by earlier versions of GATE.
Old datastores will still be readable, but if you process a corpus in an
existing datastore with this version of GATE it will become unreadable in older
versions as the documents are written back to the store after processing.

Similarly, any newly saved .gapp files that include documents will not work in
older versions of GATE without hand editing to remove the mimeType feature.