In the GUI, Save as XML (i.e. GATE format XML) now writes directly to the file stream using DocumentStaxUtils, rather than building the XML representation in memory as a String and then writing that.  This should be both faster and less memory intensive.
DocumentImpl.toXml now also uses DocumentStaxUtils writing to a StringWriter to
build the XML string (though if the XML is going to go to a file it's more
efficient to use the direct-to-file approach as above).

The XML produced by DocumentStaxUtils is compatible with that produced by the
old StringBuffer approach, the only difference being that the Node elements
within TextWithNodes are now written as <Node id="12345" /> rather than <Node
id="12345"/>, i.e. with an extra space before the slash.  This means that
round-trip loading and saving of an existing XML document will not produce the
same file byte for byte, but the result will be semantically the same.