Added default value handling for FeatureMap-valued parameters.  You can now use a creole.xml definition like:
<PARAMETER NAME="constraints" DEFAULT="kind=word;orth=upperInitial"
  COMMENT="Constraints on the annotations">gate.FeatureMap</PARAMETER>

which behaves as you would expect - the DEFAULT is treated as a list of name=value pairs separated by semicolons.  Everything up to the first equal sign is taken as the feature name, everything after is the value so

DEFAULT="queryString=id=xyz&amp;type=15;method=POST"

defines *two* features, one named "queryString" with value "id=xyz&type=15" and
another named "method" with value "POST".  Note that the ; of &amp; is part
of the entity reference and is comsumed by the XML parser, so does not act as a
separator.