A method to perform more intelligent lookup of a suitable constructor to construct an instance of one class from an instance of another.  For example, if you ask for a constructor of java.util.ArrayList that can take a java.util.LinkedList, the method will return the constructor ArrayList(Collection); if you ask for the constructor of java.util.TreeSet that takes another TreeSet it will return TreeSet(SortedSet).
This follows the rules for method lookup used by the java compiler when
compiling expressions like "List l = new ArrayList(aLinkedList)", namely it
finds the 'most specific' constructor that can be applied to the argument type.
If there is more than one applicable constructor, and no single one of them is
more specific than all the others, then an exception is thrown (this case would
cause a compile error if you tried to do it in Java code).

The test case gives some examples which should make this clearer.

For a full discussion of the method lookup algorithm and what it means for one
method to be more specific than another, see the Java Language Specification,
2nd Ed., section 15.12.2.