Modified setParameterValue so that, in the case where the specified value is not directly assignable to the parameter's type, it uses Tools.getMostSpecificConstructor to find a constructor that will instantiate the parameter's type from the passed in value.
Previously we just looked up a constructor that takes precisely the runtime
type of the specified value object, which failed in such cases as a parameter
of type ArrayList and a value generated by Arrays.asList().  There is a
suitable ArrayList constructor available (taking a Collection) but the previous
code would not find it.