Skeleton implementation of the annotation parser.  For simplicity and backwards-compatibility with creole.xml files, we will directly translate the J5 annotations into their corresponding creole.xml entries and then use the existing SAX parser.
We first load the creole.xml file into a JDOM tree, then extract all the JAR
elements and add them to the GATE classloader.  Once the classes are
accessible, we can load each class that has a RESOURCE element in creole.xml
and process its annotations, adding the relevant sub-elements to the RESOURCE.
The augmented JDOM tree is then passed to a SAXOutputter and the original
CreoleXmlHandler.

The CreoleAnnotationHandler is currently a no-op.