Added support for "resource duplication".
In a multithreaded application we often want to create several identical copies
of a resource (typically a PR or controller).  The usual way to do this is to
load a .gapp file several times, but this can be rather inefficient for certain
resources.  E.g. to create multiple copies of a DefaultGazetteer it is much
more efficient to create one in the normal way and then use several
SharedDefaultGazetteer instances to share the internal (read-only) state.
Furthermore, for some resources it is actually unsafe to create more than one
instance from the same set of parameters at the same time.

This revision introduces a Factory.duplicate(Resource) method which is intended
to return a copy of the specified resource with the same behaviour as the
original.  By default this simply calls createResource passing the same class
name, parameters, features and name as the original, but the mechanism provides
a hook for resource types that know of a better way to duplicate themselves to
override this behaviour by implementing an additional interface.