Changed Factory.deleteResource to be idempotent (i.e. to do nothing if asked to delete a resource that has already been deleted).  Previously it would call the resource's cleanup method regardless.
This required a change to the return type of ResourceData.removeInstantiation
(to return boolean rather than void).  This change is source-compatible but not
binary-compatible.  Hopefully this won't be an issue as the only place that
should have been calling ResourceData.removeInstantiation is the Factory
itself... but if anyone else does call removeInstantiation then they will need
to recompile their code for it to continue working.