Improvements to the <packagegapp> task (and thus to the export for teamware function).
The task can now cope properly with resource paths in a gapp file that point to
directories rather than files.

The task has been modified to process plugins and resource paths in
"lexicographic" order, so that in particular we now process references to a
parent directory before references to files contained under that directory.
When processing a path where we already have a mapping for one of its ancestor
directories, we can now map this path relative to its already-mapped ancestor.
This means that in cases where we have plugins contained inside other plugins
or (in onUnresolved="recover" mode) unresolved resources at different places in
the same hierarchy of directories, we maintain their relative locations in the
packaged structure.

This is particularly useful when copying the complete parent directories for
referenced resource files, as it reduces duplication in the final packaged
structure.  Previously, with unresolved references to foo/file1.jape and
foo/bar/file2.def, we would have mapped these to
application-resources/foo/file1.jape and application-resources/bar/file2.def.
But when copying the complete resource directories this would mean we have two
copies of bar in the final package (app-res/bar and app-res/foo/bar).  The new
logic simply maps foo/bar/file2.def to app-res/foo/bar/file2.def rather than
copying the directory again.