Added a concept of "priority" to @CreoleParameter annotations.  This means that when using disjunctive parameters (e.g. sourceURL vs. stringContent in DocumentImpl or grammarURL vs. binaryGrammarURL in the JAPE Transducer) specified using annotations, it is now possible to specify reliably which parameter should be the "default" one.
Previously it was up to the vagaries of the Java compiler and ASM as to which
one would be processed first, meaning that GATE compiled on one machine might
offer grammarURL by default and let you choose binaryGrammarURL from the
drop-down box, whereas the same GATE compiled on a different machine might
offer binaryGrammarURL by default and force you to select grammarURL manually.

I've also updated the places where we define disjunctive parameters in GATE
(DocumentImpl, Transducer and OWLIMOntology) to specify priorities explicitly.