My latest "on the train" project: Sharable properties.
This is an attempt to simplify the most common case where you might want to
implement CustomDuplication for a resource, in order to share objects (such as
a thread-safe cache, a pre-built FSM, etc.) between a resource and its
duplicates.  It adds a Java annotation @Sharable that can be applied to
JavaBean setter methods in the same way as for @CreoleParameter to mark that
property as sharable.  The default duplication algorithm adds all sharable
properties of the resource being duplicated to the initParams FeatureMap in the
createResource call, thus copying (by reference) the property values to the new
resource before init() is called.  Sharable properties may also be parameters
(in which case they are exempted from the usual recursive duplication) but they
need not be, and in most cases won't be as they're internal state rather than
configuration.