Long-standing bug in RepositioningInfo
Suppose you have a repositioning info like

- original offset: 5
- original length: 4
- extracted offset: 1
- extracted length: 1

- original offset: 9
- original length: 3
- extracted offset: 2
- extracted length: 3

(i.e. mapping a sequence of characters in the original content to a *shorter*
sequence of characters in the extracted content) and you ask for the extracted
offset corresponding to original offset 9 then you would previously get 5
instead of 2.  The problem is that any position within the span of the first
repos record (5-9) would be mapped to that record's extracted start offset plus
the number of excess characters in the *original* text, even if that would
return a result greater than extractedOffset+extractedLength.  I've now capped
the result so that cases like this will never return a value outside the
*extracted* span of the relevant repositioning record (any original position
greater than originalOffset+extractedLength is mapped to
extractedOffset+extractedLength).