Edge case in RepositioningInfo when the "original" content is shorter than the mapped "extracted" content.
Added a special case so that when we ask for the extracted offset that corresponds to the right hand boundary of a repositioning record we always get the right hand boundary point of the corresponding extracted segment.  This will not make a difference in most cases, but will improve accuracy when dealing with a segment where the "original" content is shorter than its corresponding "extracted" content.  In this case, previously the extracted offset returned would have been the start of the extracted segment plus the length of the *original* one, whereas now it will be the start plus the length of the extracted segment.